refactor(all): Delete unnecessary new line
불필요한 개행 제거
