+++ src/test/java/racingcar/domain/car/CarFactoryTest.java
@@ -31,5 +31,4 @@ class CarFactoryTest {
         assertThatThrownBy(() -> carFactory.makeCars(carNames))
                 .isInstanceOf(CarDuplicateException.class);
     }
-
 }
