+++ src/test/java/racingcar/interfaces/racing/RandomRacingStrategyTest.java
@@ -14,5 +14,4 @@ class RandomRacingStrategyTest {
         FakeRacingStrategy fakeRacingStrategy = new FakeRacingStrategy();
         assertThat(fakeRacingStrategy.determineCarToForward()).isEqualTo(true);
     }
-
 }
